tests/wpt/web-platform-tests/html/canvas/element/the-canvas-state/2d.state.saverestore.textBaseline.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<title>Canvas test: 2d.state.saverestore.textBaseline</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>
<link rel="stylesheet" href="/html/canvas/resources/canvas-tests.css">
<body class="show_output">

<h1>2d.state.saverestore.textBaseline</h1>
<p class="desc">save()/restore() works for textBaseline</p>


<p class="output">Actual output:</p>
<canvas id="c" class="output" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>

<ul id="d"></ul>
<script>
var t = async_test("save()/restore() works for textBaseline");
_addTest(function(canvas, ctx) {

// Test that restore() undoes any modifications
var old = ctx.textBaseline;
ctx.save();
ctx.textBaseline = "bottom";
ctx.restore();
_assertSame(ctx.textBaseline, old, "ctx.textBaseline", "old");

// Also test that save() doesn't modify the values
ctx.textBaseline = "bottom";
old = ctx.textBaseline;
    // we're not interested in failures caused by get(set(x)) != x (e.g.
    // from rounding), so compare against 'old' instead of against "bottom"
ctx.save();
_assertSame(ctx.textBaseline, old, "ctx.textBaseline", "old");
ctx.restore();


});
</script>
